refactor: Move Digitization from Plugins to Core (#1138)
This PR moves the SP builders and associated files from Plugins/Digitization to Core.
PR #1055 is split into two: reorganization of files (this PR) and updating the SP builder (will be opened shortly).
The whole Plugins/Digitization is moved to Core because the current SP builders depend on all files in Digitization, such as the DigitizationModule, PlannarModuleCluster, etc.
