+++ Core/src/Digitization/CMakeLists.txt
@@ -1,0 +1,7 @@
+target_sources(
+  ActsCore
+  PRIVATE
+    CartesianSegmentation.cpp
+    DigitizationModule.cpp
+    PlanarModuleStepper.cpp
+)
