+++ Examples/Detectors/ContextualDetector/CMakeLists.txt
@@ -9,7 +9,7 @@ target_include_directories(
 target_link_libraries(
   ActsExamplesDetectorContextual
   PUBLIC
-    ActsCore ActsPluginIdentification ActsPluginDigitization
+    ActsCore ActsPluginIdentification
     ActsExamplesFramework ActsExamplesDetectorsCommon ActsExamplesDetectorGeneric)
 
 install(
