+++ Examples/Detectors/DD4hepDetector/CMakeLists.txt
@@ -34,7 +34,7 @@ target_include_directories(
 target_link_libraries(
   ActsExamplesDetectorDD4hep
   PUBLIC
-    ActsCore ActsPluginDigitization ActsPluginDD4hep
+    ActsCore ActsPluginDD4hep
     ActsExamplesFramework ActsExamplesDetectorsCommon
     ROOT::Geom ROOT::GenVector)
