+++ Examples/Detectors/DD4hepDetector/include/ActsExamples/DD4hepDetector/DD4hepDetectorHelper.hpp
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "Acts/Plugins/Digitization/DigitizationModule.hpp"
+#include "Acts/Digitization/DigitizationModule.hpp"
 
 #include "DD4hep/Segmentations.h"
