+++ Examples/Detectors/DD4hepDetector/src/DetUtils.h
@@ -1,8 +1,8 @@
 #pragma once
 
 // DD4hep
-#include "Acts/Plugins/Digitization/CartesianSegmentation.hpp"
-#include "Acts/Plugins/Digitization/DigitizationModule.hpp"
+#include "Acts/Digitization/CartesianSegmentation.hpp"
+#include "Acts/Digitization/DigitizationModule.hpp"
 #include "DD4hep/DetFactoryHelper.h"
 #include "DD4hep/Detector.h"
 #include "DD4hep/Objects.h"
