+++ Tests/UnitTests/Core/CMakeLists.txt
@@ -2,12 +2,14 @@ add_unittest(Version VersionTests.cpp)
 
 add_subdirectory(Clusterization)
 add_subdirectory(Definitions)
+add_subdirectory(Digitization)
 add_subdirectory(EventData)
 add_subdirectory(Geometry)
 add_subdirectory(MagneticField)
 add_subdirectory(Material)
 add_subdirectory(Propagator)
 add_subdirectory(Seeding)
+add_subdirectory(SpacePointFormation)
 add_subdirectory(Surfaces)
 add_subdirectory(TrackFinding)
 add_subdirectory(TrackFitting)
