+++ Tests/UnitTests/Core/Digitization/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_unittest(CartesianSegmentation CartesianSegmentationTests.cpp)
+add_unittest(PlanarModuleStepper PlanarModuleStepperTests.cpp)
+
+
