+++ Tests/UnitTests/Core/SpacePointFormation/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_unittest(SingleHitSpacePointBuilder SingleHitSpacePointBuilderTests.cpp)
+add_unittest(DoubleHitSpacePointBuilder DoubleHitSpacePointBuilderTests.cpp)
