+++ Tests/UnitTests/Core/SpacePointFormation/DoubleHitSpacePointBuilderTests.cpp
@@ -9,9 +9,9 @@
 #include <boost/test/data/test_case.hpp>
 #include <boost/test/unit_test.hpp>
 
+#include "Acts//Digitization/DoubleHitSpacePointBuilder.hpp"
 #include "Acts/Definitions/Algebra.hpp"
-#include "Acts/Plugins/Digitization/DoubleHitSpacePointBuilder.hpp"
-#include "Acts/Plugins/Digitization/PlanarModuleCluster.hpp"
+#include "Acts/Digitization/PlanarModuleCluster.hpp"
 #include "Acts/Surfaces/PlaneSurface.hpp"
 #include "Acts/Surfaces/RectangleBounds.hpp"
 #include "Acts/Tests/CommonHelpers/DetectorElementStub.hpp"
